+++ examples/Example/ManipulateBufferLogger/ExampleWithBypassBuffer.php
@@ -6,9 +6,8 @@
 
 namespace Example\ManipulateBufferLogger;
 
-use Net\Bazzline\Component\ProxyLogger\BufferManipulator\BypassBuffer;
-use Net\Bazzline\Component\ProxyLogger\Proxy\ManipulateBufferLogger;
 use Net\Bazzline\Component\ProxyLogger\Factory\ManipulateBufferLoggerFactory;
+use Net\Bazzline\Component\ProxyLogger\Factory\BypassBufferFactory;
 use Net\Bazzline\Component\ProxyLogger\OutputToConsoleLogger;
 
 require_once __DIR__ . '/../../../vendor/autoload.php';
@@ -50,7 +49,8 @@ class ExampleWithBypassBuffer
      */
     public function setup()
     {
-        $bypassBuffer = new BypassBuffer();
+        $bypassBufferFactory = new BypassBufferFactory();
+        $bypassBuffer = $bypassBufferFactory->create();
         $logger = new OutputToConsoleLogger();
         $manipulateBufferLoggerFactory = new ManipulateBufferLoggerFactory();
         $manipulateBufferLoggerFactory->setBypassBuffer($bypassBuffer);