+++ examples/Example/ManipulateBufferLogger/ExampleWithBypassBufferAndUpwardFlushBufferTrigger.php
@@ -6,10 +6,9 @@
 
 namespace Example\ManipulateBufferLogger;
 
-use Net\Bazzline\Component\ProxyLogger\BufferManipulator\BypassBuffer;
-use Net\Bazzline\Component\ProxyLogger\BufferManipulator\UpwardFlushBufferTrigger;
-use Net\Bazzline\Component\ProxyLogger\Proxy\ManipulateBufferLogger;
+use Net\Bazzline\Component\ProxyLogger\Factory\BypassBufferFactory;
 use Net\Bazzline\Component\ProxyLogger\Factory\ManipulateBufferLoggerFactory;
+use Net\Bazzline\Component\ProxyLogger\Factory\UpwardFlushBufferTriggerFactory;
 use Net\Bazzline\Component\ProxyLogger\OutputToConsoleLogger;
 
 require_once __DIR__ . '/../../../vendor/autoload.php';
@@ -51,10 +50,12 @@ class ExampleWithBypassBufferAndUpwardFlushBufferTrigger
      */
     public function setup()
     {
-        $bypassBuffer = new BypassBuffer();
+        $bypassBufferFactory = new BypassBufferFactory();
+        $bypassBuffer = $bypassBufferFactory->create();
         $logger = new OutputToConsoleLogger();
         $manipulateBufferLoggerFactory = new ManipulateBufferLoggerFactory();
-        $flushBufferTrigger = new UpwardFlushBufferTrigger();
+        $flushBufferTriggerFactory = new UpwardFlushBufferTriggerFactory();
+        $flushBufferTrigger = $flushBufferTriggerFactory->create();
         $manipulateBufferLoggerFactory->setFlushBufferTrigger($flushBufferTrigger);
         $manipulateBufferLoggerFactory->setBypassBuffer($bypassBuffer);